ve: stable version of the VE algorithm
* Removed Factor.toBigDecimal() method, since it is no longer used.	
* Added the method Factor.isSubFactorOf().
* Changed the method FactorOperation.divide().
* Changed the method
VariableEliminationAlgorithm.calculateNormalizingConstant()
* Added tests to check the method FactorOperation.divide().
* Added tests to check the method Factor.isSubFactorOf().
* Added the "family-out" problem from [Charniak,1991].

The algorithm was not calculating the normalizing constant correctly.
The previous version considered the constant as a real number, but in
fact the normalizing constant may be a factor. All changes address the
new calculation of the constant. I have added the famous example from
[Charniak,1991] to test the algorithm. The results were manually
compared with the ones given by the JavaBayes system.