fove: get ground instance
* Added ParameterizedRandomVariable.getGroundInstance(int)
* Documentation of Constant
* Added a Population to the LogicalVariables
* Transformed Population to a immutable class
* Added PredicateSymbol.getRange()
* Added new constructor to RandomVariable
* Changed RandomVariable.toString() to o more compact representation
* Re-wrote tests for ParameterizedRandomVariable

All changes made intend to support the method
ParameterizedRandomVariable.getGroundInstance(int). 
For now, it is only possible to get a ground instance of a
parameterized random variable (PRV) if all terms in it are logical
variables. In theory, every PRV has a ground instance, but I'm not
sure if it will be necessary to get ground instances of "mixed" PRVs.

TODO
* Make all classes final
* Re-write tests for Binding, Constant, LogicalVariable and
Substitution
* Remove old constructor of RandomVariable