fove: Parameterized Factors
* Renamed packages
* Added new classes: MacroOperations, ParameterizedFactor, PRV,
ParameterizedFactorTest
* Renamed class Parfactors to Operations

Parameterized factors are factors that can hold parameterized random
variables. In terms of functionality, Parameterized Factors are the
same as Factors. I decided to put it into a separate class because
they are conceptually different. The same tests routines applied to
Factor were also written to Parameterized Factors. 
There is also a new class, PRV, that can be used to facilitate new
tests by "abbreviating" the creation of complex objects such as
parameterized random variables.
The class MacroOperations is not implemented yet.