AC-FOVE: calculation of factor value for sum out
* Created interface ParametricFactor
* Parfactor now implements interface ParametricFactor
* Filled class Operator
* Added method getFactorValue() to AggregationParfactor
* Added a manual test to check the method created

The interface ParametricFactor will be used in the future to build
mixed sets of Parfactors and AggregationParfactors. 
The Operator class was created to abstract the logical operator that
can be applied on aggregation. For now, it is very raw and ugly.
Finally, the method getFactorValue() will be used as an auxiliary
function to sum out aggregation parfactors.