C-FOVE: The algorithm
* CFOVE: class containing the algorithm and operation selection
heuristic
* MacroOperationos: new shatter method against a random variable set
* ParameterizedFactor: small change in toString method and temporary
fix to deal with precision of doubles
* SimpleParfactor: no changes, just code layout and a observation
* MathUtils: new method to calculate combination
* Parfactor: no changes
* CountingFormula: changes in getName() and two new methods:
getGroundSetSize and getPrv
* ParameterizedRandomVariable: changed the way to create an empty
instance
* PredicateSymbol: added new constructor that allows nameless functors
* Pool and CFOVETest: test of the C-FOVE algorithm
* RandomVariableSet: class that represents queries on C-FOVE. Very
similar (in structure) to counting formulas
* ConsoleFormatter and ConsoleLogger: logging classes; still in raw
stage

Still need to figure out a way to calculate factorial efficiently.
C-FOVE was tested using the example provided by Kisynski (2010) in his
thesis. 

There is also the need to replace Double with something better to
deal with calculations.