AC-FOVE: Conversion to parfactors
* Operator - re-factoring
* And - implementation of AND logical operator
* Or - implementation of OR logical operator
* BooleanOperator - self explaining interface
* AggregationParfactor - constructor and conversion to parfactors
* ParameterizedFactor - added method to create identity factor
* LogicalVariable - added constructor to copy the instance
* ParameterizedRandomVariable - constructor and factory to copy
instance
* AggregationParfactorTest and Pool: test of conversion to parfactor

Logical operators have been moved to a sub package of acfove. They are
not the best model to implement, because I'm forcing aggregation
parfactors to use boolean operators, instead of more generic ones.
Since CRALC does not use other types of operators (as far as I
understand), maybe this is ok.