AC-FOVE: Shattering and unification
* RandomVariableSet: reimplementation
* Marginal, AbstractMacroOperation: removed parametrization, new
methods
* AggParfactor: fixed imports, fixed Split to adapt to new data
structures
* MacroOperation: remodeled interface
* Shatter: fixed use of Marginal
* CFOVE, MacroOperations, Product, CFOVETest, Pool: kept for
reference, erase later
* Distribution: now is Iterable
* SplitResult: adapted to new classes, removed unnecessary methods,
fixed imports and constructors
* StdMarginal: remodeled class
* Prv, StdPrv, CountingFormula: new methods, classes are final now
* Histogram: now a final class
* Sets: improved implementation
* AggParfactorTest: adapted to new classes
* MacroOperationTest: adapted to new classes

The major change is the new way to deal with marginals. Instead of 
keeping track of every eliminable, it is now storing the query to be
made, which in turn is unlikely to change throughout the execution of
the AC-FOVE algorithm.